Minor refactor
Rename all the interfaces to give them a more sensible name
Rename the Set<InterfaceName> extension methods to Import<InterfaceName>, since evenI got confused about it's purpose for a while. It's supposed to be used by the data importer to fill in missing properties from older generation data. This means it should actually not set every property that is part of the interface and 'set' is therefor a confusing name.
Update PokeDataUI to fix layout of new property names

+++ pkNX.Structures/Personal/Info/PersonalInfo7GG.cs
@@ -51,7 +51,7 @@ public sealed class PersonalInfo7GG : IPersonalInfoGG
     public int Item2 { get => ReadInt16LittleEndian(Data.AsSpan(0x0E)); set => WriteInt16LittleEndian(Data.AsSpan(0x0E), (short)value); }
     public int Item3 { get => ReadInt16LittleEndian(Data.AsSpan(0x10)); set => WriteInt16LittleEndian(Data.AsSpan(0x10), (short)value); }
     public int Gender { get => Data[0x12]; set => Data[0x12] = (byte)value; }
-    public int HatchCycles { get => Data[0x13]; set => Data[0x13] = (byte)value; }
+    public byte HatchCycles { get => Data[0x13]; set => Data[0x13] = value; }
     public int BaseFriendship { get => Data[0x14]; set => Data[0x14] = (byte)value; }
     public int EXPGrowth { get => Data[0x15]; set => Data[0x15] = (byte)value; }
     public int EggGroup1 { get => Data[0x16]; set => Data[0x16] = (byte)value; }
